restructure files, adding libray for rating stars
