+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-rating-stars-component": "^2.2.0",
     "react-router-dom": "^6.9.0"
   },
   "devDependencies": {
