+++ src/components/RecipeDetailPageComponents/RecipeDetailSection/RecipeDescription.jsx
@@ -1,0 +1,21 @@
+
+import styles from '../../../styles/RecipeDetailPage/DetailRecipePage.module.css'
+
+const RecipeDescription = () => {
+    return (
+      <div className={styles.recipePrimaryContainer}>
+        <div className={styles.title}>Description</div>
+        <p>
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
+          tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim
+          veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea
+          commodo consequat. Duis aute irure dolor in reprehenderit in voluptate
+          velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
+          occaecat cupidatat non proident, sunt in culpa qui officia deserunt
+          mollit anim id est laborum.
+        </p>
+      </div>
+    )
+  }
+
+  export default RecipeDescription;
