+++ src/components/RecipeDetailPageComponents/RecipeReviewSection/CurrentUserReview.jsx
@@ -1,0 +1,14 @@
+
+import styles from '../../../styles/RecipeDetailPage/DetailRecipePage.module.css'
+
+
+const CurrentUserReview = () => {
+    return(
+        <div className={`${styles.currentUserReviewContainer} ${styles.flexRow}`}>
+            <div className={`${styles.starContainer} ${styles.flexItemCenter}`}></div>
+            <></>
+        </div>
+    );
+}
+
+export default CurrentUserReview;
