+++ src/types/react-rating-stars-component.d.ts
@@ -1,0 +1,1 @@
+declare module 'react-rating-stars-component';
